exekall: move NoValue in _utils.py
Remove warts trying to avoid circular dependencies for no good reasons.
